Completed CMakeLists files so that the thermoshat firmware compiles and links. Did not flash and tested it though.

+++ firmware/kw41_startup/heap_init.c
@@ -11,6 +11,6 @@
 #include "FreeRTOS.h"
 #include <stdlib.h>
 
-extern uint32_t __HeapBase;
+extern uint8_t __HeapBase;
 
 uint8_t *ucHeap = &__HeapBase;
